Fixed software metric: code duplication. Line and Projectile had exactly the same code to calculate collision, an abstract class: AbstractLine has been made to solve the issue. <organisation
